hack/services/etl_task_manager_service/requirements.txt
django
djangorestframework
drf-yasg
requests